Development

+++ src/main/java/com/wiktorski/mybudget/service/UserService.java
@@ -1,33 +1,31 @@
 package com.wiktorski.mybudget.service;
 
 
-import com.wiktorski.mybudget.model.Category;
-import com.wiktorski.mybudget.model.Payment;
-import com.wiktorski.mybudget.model.User;
+import com.wiktorski.mybudget.model.DTO.ChangePasswordRequest;
+import com.wiktorski.mybudget.model.entity.User;
 import com.wiktorski.mybudget.repository.UserRepository;
 import com.wiktorski.mybudget.service.security.SecurityService;
 import lombok.RequiredArgsConstructor;
-import org.hibernate.SessionFactory;
-import org.springframework.data.jpa.provider.HibernateUtils;
+import org.springframework.mail.SimpleMailMessage;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import java.text.SimpleDateFormat;
-import java.util.Comparator;
-import java.util.Date;
-import java.util.List;
+import javax.servlet.http.HttpServletRequest;
+import javax.validation.constraints.Email;
+import java.util.Optional;
+import java.util.UUID;
 
-//public class UserServiceImpl implements UserDetailsService {
 @Service
 @RequiredArgsConstructor
-public class UserService {
+public class UserService extends AbstractBasicService {
+
     private final UserRepository userRepo;
     private final BCryptPasswordEncoder encoder;
     private final SecurityService securityService;
+    private final EmailService emailService;
 
-    public void save(User user) {
+    public void addUser(User user) {
         user.setPassword(encoder.encode(user.getPassword()));
-        //user.setRoles(new HashSet<>(roleRepository.findAll())); // daje tu wszystkie role z bazy
         userRepo.save(user);
     }
 
@@ -39,40 +37,40 @@ public class UserService {
         return userRepo.findByConfirmationToken(token);
     }
 
-    public User findByEmail(String email) {
+    public Optional<User> findByEmail(String email) {
         return userRepo.findByEmail(email);
     }
 
-    public List<Payment> getUserPaymentsDesc() {
-        List<Payment> payments = securityService.getLoggedInUser().getPayments();
-        for (Payment p : payments) {
-            Date date = p.getDate();
-            p.setJustDate(new SimpleDateFormat("dd-MM-yyyy").format(date));
-            p.setJustTime(new SimpleDateFormat("HH:mm").format(date));
+    public void deleteUser(String token) {
+        if (token.equals(getCurrentUser().getConfirmationToken())) {
+            userRepo.delete(securityService.getLoggedInUser());
         }
-        return sortPayments(payments);
+        ;
     }
 
-    public List<Category> getUserCategories() {
-        return securityService.getLoggedInUser().getCategories();
+    public String generateConfirmationToken() {
+        User user = getCurrentUser();
+        String token = UUID.randomUUID().toString();
+        user.setConfirmationToken(token);
+        userRepo.save(user);
+        return token;
     }
 
-    private List<Payment> sortPayments(List<Payment> payments) {
-        Comparator<Payment> comp = (o1, o2) -> {
-            if (o1.getDate().before(o2.getDate())) {
-                return 1;
-            } else if (o1.getDate().after(o2.getDate())) {
-                return -1;
-            } else {
-                return 0;
-            }
-        };
-        payments.sort(comp);
-        return payments;
+    /******************************************************************************************/
+    //TODO those methods should be in user entity as there are stored those informations
+    public void addToBudget(float amount) {
+        User user = securityService.getLoggedInUser();
+        user.setBudget(user.getBudget() + amount);
     }
 
-    public void deleteUser() {
-        userRepo.delete(securityService.getLoggedInUser());
+    public void decreaseBudget(float price) {
+        User user = securityService.getLoggedInUser();
+        user.setBudget(user.getBudget() - price);
+    }
+
+    public void decreaseBudget(float price, User user) {
+        user.setBudget(user.getBudget() - price);
+        userRepo.save(user);
     }
 
     public void setBudget(float amount) {
@@ -80,18 +78,21 @@ public class UserService {
         user.setBudget(amount);
     }
 
-    public void addToBudget(float amount){
-        User user = securityService.getLoggedInUser();
-        float tmp=user.getBudget()+amount;
-        user.setBudget(tmp);
+    public void sendDeleteConfirmationEmail(HttpServletRequest request) {
+        generateConfirmationToken();
+        SimpleMailMessage deleteConfirmationEmail = emailService.createDeleteConfirmationEmail(getCurrentUser(), request);
+        emailService.sendEmail(deleteConfirmationEmail);
     }
 
-    //decrease budget amount
-    public void addPaymentToBudget(float price){
-        User user = securityService.getLoggedInUser();
-        user.setBudget(user.getBudget()-price);
+    public boolean changePassword(ChangePasswordRequest req) {
+        User user = getCurrentUser();
+        if (encoder.matches(req.getCurrentPassword(), user.getPassword())) {
+            user.setPassword(encoder.encode(req.getNewPassword()));
+            userRepo.save(user);
+            return true;
+        }
+        return false;
     }
-
 
 
     /*@Override
